pyproject.toml
[tool.poetry]
name = "dagpi-image"
version = "1.0.0"
description = ""
authors = ["Daggy1234 <arnav.jindal7@gmail.com>"]
license = "AGPLv3"

[tool.poetry.dependencies]
python = "^3.8"
Wand = "0.6.5"
httptools = "0.1.1"
async-timeout = "3.0.1"
httpx = "0.16.1"
starlette-prometheus = "0.7.0"
sentry-sdk = "0.18.0"
scipy = "1.5.2"
Pillow = "8.1.1"
uvicorn = "0.12.1"
fastapi = "0.63.0"
matplotlib = "3.3.4"
scikit-image = "0.17.2"
numpy = "1.20.1"
gunicorn = "^20.0.4"
polaroid = "0.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
